// src/app/Model/production.ts
export class ProductionLineCode {
  WIPProcessCode: string;
  WIPProcessName: string;
}
// ...
export class ProcessData {
  WIPProcessCode: string = "";
  WIPProcessType: string = "";
  WIPProcessName: string = "";
  WIPProcessStatus:string = "";
  ProductionLineCode: string = "";
  ProductionLineName: string = "";
  PartNo:string = "";
  PartDesc:string = "";
  IsActive: boolean = false;
  Quantity:number = 0;
}